+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-16 20:45:52
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-08-16 22:05:43
+// Commit: test(theme): implement widget integration tests for theme
 
 class themeComponentState {
-  final String commitRef = '3657';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-16T20:45:52+05:00');
+  final String commitRef = '3660';
+  final String lastMessage = 'test(theme): implement widget integration tests for theme';
+  final DateTime timestamp = DateTime.parse('2026-08-16T22:05:43+05:00');
 }